requirement.txt
Django==5.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
firebase-admin==7.1.0
django-cors-headers==4.7.0
whitenoise==6.12.0
gunicorn==25.3.0
psycopg2-binary==2.9.10
